swarnim-survase-to-do-list/requirements.txt
Flask
flask-cors
langchain
langchain-google-genai
langchain-core
python-dotenv
pydantic             

gunicorn 
